Amplify each sound with the lowest common value, 6.542dB
In KDE chat I heard a complaint about the sounds being too quiet. When inspecting them, I have found that none of them are amplified to the max, all could be louder. What I have done is finding which sound is the loudest, and so which is needing the least amplification to maximize volume, and amplified all sounds with that value. Then I exported the sounds in Tenacity, on max quality, whence the increased size. With using the same amplification for all sounds it won't be audible, that any one sound is more loud, than it was previously. (But if need be, each sound could be amplified on it's own)

On a different note, when trying Tenacity, it turned out it does not support oga. I tried ffmpeg, which does, but when using it, it recognized the sounds as ogg. This was also confirmed by the fact that when amplifying and converting to oga, Tenacity could not open the sounds anymore, but when keeping them as ogg, it could. As a conclusion, the sounds are named oga but are of the format ogg. This commit did not change this, I kept the names as they were, but I would either recommend renaming them, or converting to oga.

I was told to ping @raploz about this.