audio-volume-change: bring the sound volume way down
approx linear -20db to be precise. that is where the freedesktop sound
peaks out as well. with all headphones I have that produces a reasonably
loud sound at 100% volume but not so loud as to be painful (as was the
case before)

file has been adjusted in audacity, exported to wav, then converted to
oga as per readme

BUG: 477125